MSK20FRONT-15 [CPM] Ticket table & ticket info
