+++ client/components/cpm/info/info.component.js
@@ -1,0 +1,18 @@
+/**
+ * Created by vlados on 28.11.2020
+ */
+import Component from '../../../modules/component/component.js';
+import store from '../../../store/index.js';
+import {template} from "./info.template.js";
+
+export default class info extends Component {
+    constructor() {
+        super({
+            store
+        });
+    }
+
+    render() {
+        return this.compile(template.call(this));
+    }
+}
