+++ client/components/cpm/info/info.template.js
@@ -1,0 +1,14 @@
+export const template = function () {
+    return {
+        tagName: 'div',
+        classList: [''],
+        attributes: {
+            style: 'color: black'
+        },
+        events: {
+            // onclick: this.methods.makeSth
+        },
+        textContent: '',
+        children: []
+    }
+}
