+++ client/less/components/cpm/card.less
@@ -1,0 +1,5 @@
+.card{
+  background: #FFFFFF;
+  border: 1px solid #DDE4EB;
+  box-shadow: 0px 0px 0px rgba(90, 113, 208, 0.107337);
+}
